Refactored card, created folder for icon components
