+++ MyWeather/src/OtherConditions.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import UVIcon from "./UVIcon";
+import UVIcon from "../src/IconComponents/UVIcon";
 
 // Icon Imports
 import WeatherVane from '../src/assets/weather_sagittarius.svg';
